==> scATAC-seq/4-clustering/1-level_1/02-normalization_and_integration_with_multiome_newpeakcalling.Rmd <==
---
title: "Normalization and Batch Effect Correction after the new Peak calling"
author: "Paula Soler-Vila"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Introduction

The new peak calling round applied on the previous notebook significantly increased the number of the features identified in our dataset. Therefore, we must need to repeat the standard downstream analysis, including data normalization, dimensionality reduction analysis and batch correction to account for this change in the number of features detected.


# Pre-processing

## Load packages

```{r}
library(Seurat)
library(SeuratWrappers)
library(Signac)
library(harmony)
library(tidyverse)

set.seed(1234)
```

## Parameters

```{r}
# Paths
path_to_obj <- here::here("scATAC-seq/results/R_objects/8.1.tonsil_atac_integrated_with_multiome_annotated_level1_new_peakcalling.rds")
path_to_save_obj <- here::here("scATAC-seq/results/R_objects/8.2.tonsil_peakcalling_annotation_level1_integrated.rds")
path_tmp_dir <- here::here("scATAC-seq/4-clustering/1-level_1/tmp/")

path_to_save_dimred_uncorrect <- str_c(path_tmp_dir, "batch_uncorrected_lsi.rds", sep = "")
path_to_save_dimred_correct <- str_c(path_tmp_dir, "batch_corrected_lsi.rds", sep = "")
path_to_save_confounders_df <- str_c(path_tmp_dir, "confounders_df.rds", sep = "") 
```

## Load data

```{r}
tonsil <- readRDS(path_to_obj)
tonsil
```

# Visualize UMAP without batch effect correction

```{r fig.wide=TRUE}
# Process Seurat object
tonsil <- tonsil %>%
  RunTFIDF() %>% 
  FindTopFeatures(min.cutoff = "q0") %>%
  RunSVD() %>%
  RunUMAP(reduction = "lsi", dims = 2:40)

DepthCor(tonsil)

# Visualize UMAP
confounders <- c("library_name", "sex", "age_group", "hospital", "assay")
umaps_before_integration <- purrr::map(confounders, function(x) {
  p <- DimPlot(tonsil, group.by = x, pt.size = 0.1)
  p
})
names(umaps_before_integration) <- confounders
print("UMAP colored by GEM:")
umaps_before_integration$library_name + NoLegend()
print("UMAP colored by sex, age group, cell hashing status, sampling center and assay:")
umaps_before_integration[2:length(umaps_before_integration)]
```


# Run and visualize Harmony's integration

```{r fig.wide=TRUE}
tonsil <- RunHarmony(
  object = tonsil,
  group.by.vars = "gem_id", 
  reduction = "lsi",
  dims = 2:40,
  assay.use = "peaks_macs",
  project.dim = FALSE
)

tonsil <- RunUMAP(tonsil, dims = 2:40, reduction = "harmony")

umaps_after_integration <- purrr::map(confounders, function(x) {
  p <- DimPlot(tonsil, group.by = x, pt.size = 0.1)
  p
})
names(umaps_after_integration) <- confounders
print("UMAP colored by GEM:")
umaps_after_integration$library_name + NoLegend()
print("UMAP colored by sex, age group, cell hashing status, sampling center and assay:")
umaps_after_integration[2:length(umaps_before_integration)]
```

# Save

```{r}
# Save integrated Seurat object
saveRDS(tonsil, path_to_save_obj)


# Save PCA matrices to compute the Local Inverse Simpson Index (LISI)
confounders_df <- tonsil@meta.data[, confounders]
saveRDS(confounders_df, path_to_save_confounders_df)
saveRDS(
  tonsil@reductions$lsi@cell.embeddings[, 2:40],
  path_to_save_dimred_uncorrect
)
saveRDS(
  tonsil@reductions$harmony@cell.embeddings[, 2:40],
  path_to_save_dimred_correct
)
```


# Session Information

```{r}
sessionInfo()
```
